# swagger_server/swagger/swagger.yaml
openapi: 3.0.1
info:
  title: backend
  version: 0.0.1
servers:
- url: http://localhost:8080/api
  description: Development server
- url: https://localhost:8080/api
  description: Development server with TLS Profile
security:
- bearerAuth: []
...
paths:
  /auth/send_verification_code:
    post:
      tags:
      - auth
      description: Send SMS with verification code
      operationId: auth_send_verification_code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerificationCodeRequest'
      responses:
        default:
          description: OK
          content: {}
      security: []
      x-openapi-router-controller: swagger_server.controllers.auth_controller
  /auth/verify_phone_code:
    post:
      tags:
      - auth
      description: Verify the code sent via SMS
      operationId: auth_verify_phone_code
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyPhoneCodeRequest'
      responses:
        "400":
          description: Verification failed
        default:
          description: Basic response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseBase'
      security: []
      x-openapi-router-controller: swagger_server.controllers.auth_controller
  /auth/sign_up:
    post:
      tags:
      - auth
      description: Adds a new user to the system if they don’t exist yet
      operationId: auth_sign_up
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignUpRequest'
      responses:
        default:
          description: User registered and automatically logged in
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SignUpResponse'
      security: []
      x-openapi-router-controller: swagger_server.controllers.auth_controller
...
components:
  schemas:
    UserId:
      type: integer
...
    Phone:
      pattern: \+[0-9]{7,20}
      type: string
      description: Phone number. Stripped from any punctuation except for leading
        '+'.
    VerificationCode:
      pattern: '[0-9]{4}'
      type: string
      description: Verification code from SMS. 4-digits string.
    AgeGroup:
      type: string
      enum:
      - <18
      - 18-25
      - 26-35
      - 36-45
      - '>45'
    LanguageSpec:
      type: string
      description: https://docs.oracle.com/javase/tutorial/i18n/locale/matching.html
      enum:
      - en-US
      - ru-RU
      - fr-FR
      - it-IT
      - es-ES
    ResponseBase:
      required:
      - status
      type: object
      properties:
        status:
          type: integer
        error_message:
          type: string
      example:
        error_message: error_message
        status: 0
    Image:
      type: string
      description: Base64 encoded JPEG image
      format: byte
...
    UserInfo:
      required:
      - age_group
      - id
      - image
      - language
      - name
      type: object
      properties:
        id:
          $ref: '#/components/schemas/UserId'
        name:
          type: string
        age_group:
          $ref: '#/components/schemas/AgeGroup'
        language:
          $ref: '#/components/schemas/LanguageSpec'
        image:
          $ref: '#/components/schemas/Image'
      example:
        image: ""
        age_group: <18
        name: name
        language: en-US
        id: 6
    UserInfoSelf:
      required:
      - private
      - public
      type: object
      properties:
        public:
          $ref: '#/components/schemas/UserInfo'
        private:
          $ref: '#/components/schemas/UserInfoSelf_private'
      example:
        private:
          phone: phone
          token: token
        public:
          image: ""
          age_group: <18
          name: name
          language: en-US
          id: 6
    VerificationCodeRequest:
      required:
      - phone
      type: object
      properties:
        phone:
          $ref: '#/components/schemas/Phone'
    VerifyPhoneCodeRequest:
      required:
      - code
      - phone
      type: object
      properties:
        phone:
          $ref: '#/components/schemas/Phone'
        code:
          $ref: '#/components/schemas/VerificationCode'
    SignUpRequest:
      required:
      - age_group
      - code
      - language
      - name
      - phone
      type: object
      properties:
        phone:
          $ref: '#/components/schemas/Phone'
        code:
          $ref: '#/components/schemas/VerificationCode'
        name:
          type: string
        age_group:
          $ref: '#/components/schemas/AgeGroup'
        language:
          $ref: '#/components/schemas/LanguageSpec'
    SignUpResponse:
      required:
      - status
      type: object
      properties:
        status:
          type: integer
        error_message:
          type: string
        me:
          $ref: '#/components/schemas/UserInfoSelf'
      example:
        error_message: error_message
        me:
          private:
            phone: phone
            token: token
          public:
            image: ""
            age_group: <18
            name: name
            language: en-US
            id: 6
        status: 0
...
    UserInfoSelf_private:
      required:
      - phone
      - token
      type: object
      properties:
        phone:
          $ref: '#/components/schemas/Phone'
        token:
          type: string
          description: Authentication token
      example:
        phone: phone
        token: token
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      x-bearerInfoFunc: swagger_server.controllers.authorization_controller.check_bearerAuth